added kafka relationship on tweets splitted tweet routes / client promise resolution set "elasticsearch" as shared service name
