+++ .platform/services.yaml
@@ -1,4 +1,8 @@
-rssreaderelasticsearch:
+elasticsearch:
   type: elasticsearch:6.5
   disk: 1024
   size: S
+
+kafkastreaming:
+  type: kafka:2.2
+  disk: 1024
