+++ tweets/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "dotenv": "^8.0.0",
     "install": "^0.12.2",
+    "kafka-node": "^4.1.3",
     "micro": "^9.3.4",
     "node-cache": "^4.2.0",
     "npm": "^6.10.2",
